+++ xdef/src/main/java/org/xdef/impl/code/DefXmlWriter.java
@@ -21,6 +21,7 @@ import org.w3c.dom.Node;
 import org.xdef.XDContainer;
 import org.xdef.XDValueType;
 import java.math.BigInteger;
+import org.xdef.sys.CurrencyAmount;
 import org.xdef.sys.GPSPosition;
 
 /** Provides incremental writing of XML data to a data stream.
@@ -156,6 +157,8 @@ public class DefXmlWriter extends KXmlOutStream
 	public XDContainer contextValue() {return null;}
 	@Override
 	public GPSPosition GPSValue() {return null;}
+	@Override
+	public CurrencyAmount CurrencyAmount() {return null;}
 	@Override
 	public byte[] getBytes() {return null;}
 	@Override